Move warp code to routes.rs
